refactor(observability): move orchestra logging to observability module
Move vibe3.orchestra.logging to vibe3.observability.orchestra_log to
establish proper dependency boundaries. Domain layer now imports from
observability instead of orchestra infrastructure.

Changes:
- Create src/vibe3/observability/orchestra_log.py with all logging functions
- Update src/vibe3/observability/__init__.py to re-export logging symbols
- Convert src/vibe3/orchestra/logging.py to compatibility shell
- Update src/vibe3/orchestra/__init__.py to import from observability
- Update domain layer imports (5 files) to use observability.orchestra_log
- Remove orchestra.logging from allowed_orchestra_modules in architecture test
- Update test imports and mock patches to use new module location

All tests pass: domain architecture guard, logging tests, and affected modules.
No behavior changes - purely import path relocation with backward compatibility.
