+++ tests/vibe3/domain/handlers/test_supervisor_scan.py
@@ -9,7 +9,7 @@ from vibe3.execution.contracts import ExecutionLaunchResult
 class TestSupervisorScanHandler:
     """supervisor_scan handler dispatches supervisor apply via CLI self-invocation."""
 
-    @patch("vibe3.orchestra.logging.append_orchestra_event")
+    @patch("vibe3.observability.orchestra_log.append_orchestra_event")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.domain.handlers.supervisor_scan.load_orchestra_config")
@@ -110,7 +110,7 @@ class TestSupervisorScanHandler:
         mock_coordinator_cls.assert_not_called()
         mock_get_store.assert_not_called()
 
-    @patch("vibe3.orchestra.logging.append_orchestra_event")
+    @patch("vibe3.observability.orchestra_log.append_orchestra_event")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.domain.handlers.supervisor_scan.load_orchestra_config")
@@ -144,7 +144,7 @@ class TestSupervisorScanHandler:
 
         mock_coordinator.dispatch_execution.assert_called_once()
 
-    @patch("vibe3.orchestra.logging.append_orchestra_event")
+    @patch("vibe3.observability.orchestra_log.append_orchestra_event")
     @patch("vibe3.clients.sqlite_client.SQLiteClient")
     @patch("vibe3.execution.coordinator.ExecutionCoordinator")
     @patch("vibe3.domain.handlers.supervisor_scan.load_orchestra_config")
