+++ tests/vibe3/domain/test_no_orchestra_imports.py
@@ -60,7 +60,6 @@ def is_orchestra_import(import_str: str, file_path: Path | None = None) -> bool:
 
     # Allowed imports (infrastructure services)
     allowed_orchestra_modules = {
-        "vibe3.orchestra.logging",  # Infrastructure logging
         "vibe3.orchestra.protocols",  # TYPE_CHECKING only (checked by caller)
         "vibe3.orchestra.failed_gate",  # TYPE_CHECKING only (checked by caller)
         # Orchestra internal services (used via protocol injection)
@@ -148,8 +147,7 @@ def test_domain_layer_no_orchestra_imports():
             f"Found {len(violations)} orchestra imports in domain layer:\n"
             f"{violation_list}\n\n"
             "Domain layer should not import from orchestra layer.\n"
-            "Allowed exceptions: orchestra.logging (infrastructure), "
-            "orchestra.protocols/failed_gate (TYPE_CHECKING only)."
+            "Allowed exceptions: orchestra.protocols/failed_gate (TYPE_CHECKING only)."
         )
 
     print(f"✓ Verified {len(list(domain_dir.rglob('*.py')))} domain files")
